unifies upload icon
