--- app/components/uploader/index.ts
@@ -1,2 +1,0 @@
-export { default } from './component';
-export type { UploaderProps } from './types';
