--- app/components/uploader/types.ts
@@ -1,11 +1,0 @@
-import { ReactChild } from 'react';
-
-export interface UploaderProps {
-  id?: string;
-  open?: boolean;
-  caption?: string;
-  disabled?: boolean;
-  children: ReactChild | ReactChild[];
-  onOpen: () => void;
-  onClose: () => void;
-}
