+++ app/package.json
@@ -88,6 +88,7 @@
     "react-dom": "18.2.0",
     "react-dropzone": "^11.3.1",
     "react-final-form": "6.5.9",
+    "react-icons": "4.11.0",
     "react-intersection-observer": "^8.31.1",
     "react-joyride": "^2.3.0",
     "react-map-gl": "^6.1.13",
